requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Cache & Async
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.1.1
cryptography==41.0.7
PyJWT==2.8.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# External APIs
stripe==7.4.0
sendgrid==6.11.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1
